refactor: extract Sankey diagram rendering logic into useRenderDiagram hook
